requirements.txt
torch==2.2.0
torchvision==0.17.0
torchmetrics
numpy==1.26.4
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core==1.3.2
pyloudnorm
torchaudio
soundfile
pystoi

black
isort
pre-commit
